Do not extend the utility component anymore
Fix an issue on showing card image
