--- src/credit-card-utilities.js
@@ -1,407 +1,0 @@
-// @flow
-
-import React, { Component } from 'react';
-import payment from 'payment';
-import creditCardType from 'credit-card-type';
-
-import {
-  formatCardNumber,
-  formatExpiry,
-  formatCvc,
-  hasCardNumberReachedMaxLength,
-  hasCVCReachedMaxLength,
-  hasZipReachedMaxLength,
-  isHighlighted
-} from './utils/formatter';
-import images from './utils/images';
-import isExpiryInvalid from './utils/is-expiry-invalid';
-import isZipValid from './utils/is-zip-valid';
-
-const BACKSPACE_KEY_CODE = 8;
-const CARD_TYPES = {
-  mastercard: 'MASTERCARD',
-  visa: 'VISA',
-  amex: 'AMERICAN_EXPRESS'
-};
-
-type Props = {
-  CARD_TYPES: Object,
-  cardCVCInputRenderer: Function,
-  cardExpiryInputRenderer: Function,
-  cardNumberInputRenderer: Function,
-  cardZipInputRenderer: Function,
-  onError?: Function,
-  cardExpiryInputProps: Object,
-  cardNumberInputProps: Object,
-  cardCVCInputProps: Object,
-  cardZipInputProps: Object,
-  cardImageClassName: string,
-  cardImageStyle: Object,
-  containerClassName: string,
-  containerStyle: Object,
-  dangerTextClassName: string,
-  dangerTextStyle: Object,
-  fieldClassName: string,
-  fieldStyle: Object,
-  enableZipInput: boolean,
-  images: Object,
-  inputComponent: Function | Object | string,
-  inputClassName: string,
-  inputStyle: Object,
-  invalidClassName: string,
-  invalidStyle: Object,
-  customTextLabels: Object
-};
-
-type State = {
-  cardImage: string,
-  cardNumberLength: number,
-  cardNumber: ?string,
-  errorText: ?string,
-  showZip: boolean
-};
-
-const inputRenderer = ({ props }: Object) => <input {...props} />;
-
-class CreditCardUtilities extends Component<Props, State> {
-  cardExpiryField: any;
-  cardNumberField: any;
-  cvcField: any;
-  zipField: any;
-
-  static defaultProps = {
-    cardCVCInputRenderer: inputRenderer,
-    cardExpiryInputRenderer: inputRenderer,
-    cardNumberInputRenderer: inputRenderer,
-    cardZipInputRenderer: inputRenderer,
-    cardExpiryInputProps: {},
-    cardNumberInputProps: {},
-    cardCVCInputProps: {},
-    cardZipInputProps: {},
-    cardImageClassName: '',
-    cardImageStyle: {},
-    containerClassName: '',
-    containerStyle: {},
-    dangerTextClassName: '',
-    dangerTextStyle: {},
-    enableZipInput: false,
-    fieldClassName: '',
-    fieldStyle: {},
-    inputComponent: 'input',
-    inputClassName: '',
-    inputStyle: {},
-    invalidClassName: 'is-invalid',
-    invalidStyle: {},
-    customTextLabels: {}
-  };
-
-  constructor(props: Props) {
-    super(props);
-    this.CARD_TYPES = Object.assign({}, CARD_TYPES, props.CARD_TYPES);
-    this.images = Object.assign({}, images, props.images);
-    this.state = {
-      cardImage: this.images.placeholder,
-      cardNumberLength: 0,
-      cardNumber: null,
-      errorText: null,
-      showZip: false
-    };
-  }
-
-  componentDidMount = () => {
-    this.setState({ cardNumber: this.cardNumberField.value }, () => {
-      const cardType = payment.fns.cardType(this.state.cardNumber);
-      const images = this.images;
-      this.setState({
-        cardImage: images[cardType] || images.placeholder
-      });
-    });
-  };
-
-  isMonthDashKey = ({ key, target: { value } } = {}) => {
-    return !value.match(/[/-]/) && /^[/-]$/.test(key);
-  };
-
-  checkIsNumeric = (e: any) => {
-    if (!/^\d*$/.test(e.key)) {
-      e.preventDefault();
-    }
-  };
-
-  handleCardNumberBlur = (
-    { onBlur }: { onBlur?: ?Function } = { onBlur: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    if (!payment.fns.validateCardNumber(e.target.value)) {
-      this.setFieldInvalid(
-        customTextLabels.invalidCardNumber || 'Card number is invalid',
-        'cardNumber'
-      );
-    }
-
-    const { cardNumberInputProps } = this.props;
-    cardNumberInputProps.onBlur && cardNumberInputProps.onBlur(e);
-    onBlur && onBlur(e);
-  };
-
-  handleCardNumberChange = (
-    { onChange }: { onChange?: ?Function } = { onChange: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const {
-      customTextLabels,
-      enableZipInput,
-      cardNumberInputProps
-    } = this.props;
-    const images = this.images;
-    const cardNumber = e.target.value;
-    const cardNumberLength = cardNumber.split(' ').join('').length;
-    const cardType = payment.fns.cardType(cardNumber);
-    const cardTypeInfo =
-      creditCardType.getTypeInfo(
-        creditCardType.types[this.CARD_TYPES[cardType]]
-      ) || {};
-    const cardTypeLengths = cardTypeInfo.lengths || [16];
-
-    this.cardNumberField.value = formatCardNumber(cardNumber);
-
-    this.setState({
-      cardImage: images[cardType] || images.placeholder,
-      cardNumber
-    });
-
-    if (enableZipInput) {
-      this.setState({ showZip: cardNumberLength >= 6 });
-    }
-
-    this.setFieldValid();
-    if (cardTypeLengths) {
-      const lastCardTypeLength = cardTypeLengths[cardTypeLengths.length - 1];
-      for (let length of cardTypeLengths) {
-        if (
-          length === cardNumberLength &&
-          payment.fns.validateCardNumber(cardNumber)
-        ) {
-          this.cardExpiryField.focus();
-          break;
-        }
-        if (cardNumberLength === lastCardTypeLength) {
-          this.setFieldInvalid(
-            customTextLabels.invalidCardNumber || 'Card number is invalid',
-            'cardNumber'
-          );
-        }
-      }
-    }
-
-    cardNumberInputProps.onChange && cardNumberInputProps.onChange(e);
-    onChange && onChange(e);
-  };
-
-  handleCardNumberKeyPress = (e: any) => {
-    const value = e.target.value;
-    this.checkIsNumeric(e);
-    if (value && !isHighlighted()) {
-      const valueLength = value.split(' ').join('').length;
-      if (hasCardNumberReachedMaxLength(value, valueLength)) {
-        e.preventDefault();
-      }
-    }
-  };
-
-  handleCardExpiryBlur = (
-    { onBlur }: { onBlur?: ?Function } = { onBlur: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    const cardExpiry = e.target.value.split(' / ').join('/');
-    const expiryError = isExpiryInvalid(
-      cardExpiry,
-      customTextLabels.expiryError
-    );
-    if (expiryError) {
-      this.setFieldInvalid(expiryError, 'cardExpiry');
-    }
-
-    const { cardExpiryInputProps } = this.props;
-    cardExpiryInputProps.onBlur && cardExpiryInputProps.onBlur(e);
-    onBlur && onBlur(e);
-  };
-
-  handleCardExpiryChange = (
-    { onChange }: { onChange?: ?Function } = { onChange: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-
-    this.cardExpiryField.value = formatExpiry(e);
-    const value = this.cardExpiryField.value.split(' / ').join('/');
-
-    this.setFieldValid();
-
-    const expiryError = isExpiryInvalid(value, customTextLabels.expiryError);
-    if (value.length > 4) {
-      if (expiryError) {
-        this.setFieldInvalid(expiryError, 'cardExpiry');
-      } else {
-        this.cvcField.focus();
-      }
-    }
-
-    const { cardExpiryInputProps } = this.props;
-    cardExpiryInputProps.onChange && cardExpiryInputProps.onChange(e);
-    onChange && onChange(e);
-  };
-
-  handleCardExpiryKeyPress = (e: any) => {
-    const value = e.target.value;
-
-    if (!this.isMonthDashKey(e)) {
-      this.checkIsNumeric(e);
-    }
-
-    if (value && !isHighlighted()) {
-      const valueLength = value.split(' / ').join('').length;
-      if (valueLength >= 4) {
-        e.preventDefault();
-      }
-    }
-  };
-
-  handleCardCVCBlur = (
-    { onBlur }: { onBlur?: ?Function } = { onBlur: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    if (!payment.fns.validateCardCVC(e.target.value)) {
-      this.setFieldInvalid(
-        customTextLabels.invalidCvc || 'CVC is invalid',
-        'cardCVC'
-      );
-    }
-
-    const { cardCVCInputProps } = this.props;
-    cardCVCInputProps.onBlur && cardCVCInputProps.onBlur(e);
-    onBlur && onBlur(e);
-  };
-
-  handleCardCVCChange = (
-    { onChange }: { onChange?: ?Function } = { onChange: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    const value = formatCvc(e.target.value);
-    this.cvcField.value = value;
-    const CVC = value;
-    const CVCLength = CVC.length;
-    const isZipFieldAvailable = this.props.enableZipInput && this.state.showZip;
-    const cardType = payment.fns.cardType(this.state.cardNumber);
-
-    this.setFieldValid();
-    if (CVCLength >= 4) {
-      if (!payment.fns.validateCardCVC(CVC, cardType)) {
-        this.setFieldInvalid(
-          customTextLabels.invalidCvc || 'CVC is invalid',
-          'cardCVC'
-        );
-      }
-    }
-
-    if (isZipFieldAvailable && hasCVCReachedMaxLength(cardType, CVCLength)) {
-      this.zipField.focus();
-    }
-
-    const { cardCVCInputProps } = this.props;
-    cardCVCInputProps.onChange && cardCVCInputProps.onChange(e);
-    onChange && onChange(e);
-  };
-
-  handleCardCVCKeyPress = (e: any) => {
-    const cardType = payment.fns.cardType(this.state.cardNumber);
-    const value = e.target.value;
-    this.checkIsNumeric(e);
-    if (value && !isHighlighted()) {
-      const valueLength = value.split(' / ').join('').length;
-      if (hasCVCReachedMaxLength(cardType, valueLength)) {
-        e.preventDefault();
-      }
-    }
-  };
-
-  handleCardZipBlur = (
-    { onBlur }: { onBlur?: ?Function } = { onBlur: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    if (!isZipValid(e.target.value)) {
-      this.setFieldInvalid(
-        customTextLabels.invalidZipCode || 'Zip code is invalid',
-        'cardZip'
-      );
-    }
-
-    const { cardZipInputProps } = this.props;
-    cardZipInputProps.onBlur && cardZipInputProps.onBlur(e);
-    onBlur && onBlur(e);
-  };
-
-  handleCardZipChange = (
-    { onChange }: { onChange?: ?Function } = { onChange: null }
-  ) => (e: SyntheticInputEvent<*>) => {
-    const { customTextLabels } = this.props;
-    const zip = e.target.value;
-    const zipLength = zip.length;
-
-    this.setFieldValid();
-
-    if (zipLength >= 5 && !isZipValid(zip)) {
-      this.setFieldInvalid(
-        customTextLabels.invalidZipCode || 'Zip code is invalid',
-        'cardZip'
-      );
-    }
-
-    const { cardZipInputProps } = this.props;
-    cardZipInputProps.onChange && cardZipInputProps.onChange(e);
-    onChange && onChange(e);
-  };
-
-  handleCardZipKeyPress = (e: any) => {
-    const cardType = payment.fns.cardType(this.state.cardNumber);
-    const value = e.target.value;
-    this.checkIsNumeric(e);
-    if (value && !isHighlighted()) {
-      const valueLength = value.split(' / ').join('').length;
-      if (hasZipReachedMaxLength(cardType, valueLength)) {
-        e.preventDefault();
-      }
-    }
-  };
-
-  handleKeyDown = (ref: any) => {
-    return (e: SyntheticInputEvent<*>) => {
-      if (e.keyCode === BACKSPACE_KEY_CODE && !e.target.value) {
-        ref.focus();
-      }
-    };
-  };
-
-  setFieldInvalid = (errorText: string, inputName?: string) => {
-    const { invalidClassName, onError } = this.props;
-    // $FlowFixMe
-    document.getElementById('field-wrapper').classList.add(invalidClassName);
-    this.setState({ errorText });
-
-    if (inputName) {
-      const { onError } = this.props[`${inputName}InputProps`];
-      onError && onError(errorText);
-    }
-
-    if (onError) {
-      onError({ inputName, error: errorText });
-    }
-  };
-
-  setFieldValid = () => {
-    const { invalidClassName } = this.props;
-    // $FlowFixMe
-    document.getElementById('field-wrapper').classList.remove(invalidClassName);
-    this.setState({ errorText: null });
-  };
-}
-
-export default CreditCardUtilities;
